Add a .bytes() method to HTTP responses
The new .bytes() method returns the body as bytes rather than a string.
It is modeled after the .json() method, which parses the body as JSON.

The use case is parsing HTTP requests that return binary data. Certain
operations on strings in Starlark don't work correctly with binary
(i.e., non-UTF-8 encoded) data. In particular, when strings contain
non-UTF-8 binary data, the `ord` function returns 65533 (the codepoint
for the Unicode replacement character) when called on a byte with a
value between 0x80 and 0xff. `ord` functions as expected with bytes
values.